requirements.txt
numpy==2.2.6
torch==2.6.0
transformers==4.51.3
llama-index
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
langchain-openai
llama-index-llms-langchain
langchain-community
pyzmq==26.2.0
psycopg2